3_Agent_to_Deploy_Assets/cdk-python/requirements.txt
aws-cdk-lib>=2.70.0
constructs>=10.0.0